Created Header File
Separated the class from main.cpp
