--- Police Database/Untitled.h
@@ -1,7 +1,0 @@
-//
-//  Untitled.h
-//  Police Database
-//
-//  Created by Luis Herrera on 9/22/24.
-//
-
